+++ Changelog.md
@@ -1,3 +1,10 @@
+## 0.3.0 (in development)
+
+- Refactor
+- Unit Tests (Karma + Chai)
+- `track.options.looping` &rarr; `track.options.loop`
+- `track.tweenPan()` now uses `linearRampToValueAtTime()`
+
 ## 0.2.0
 
 - audio analysis